cryspy_numbers and cryspy_fromstr are working.
cryspy_numbers contains the numbers and matrix features.
It is finished (as far as I didn't forget anything) and working.
cryspy_fromstr contains the function for creating object.
Numbers and Matrices can be created already.
So far I worked with doctest, but I'm going to switch to pytest in the further modules.