cryspy_geo, cryspy_crystal
I fixed a bug in cryspy_geo: A Transformation is created by equations for the basis-vectors (in cryspy_fromstr) and is displayed in such a way. So far, this equation contained the matrix-components for the Transformation. However, the matrix-components are with respect to the coordinates of a point, not with respect to the basis vectors.
Fix: It is still created and displayed as equations of basis-vectors, but the actual matrix is inverted, so that the Transformation acts on coordinate-vectors in a proper way.

In order to make the application of a Operator on an Atom like operator**atom, I had to override Operator.__pow__ by Atom.__rpow__ . I changed the error-message of Operator.__pow__ to "return NotImplemented", so that the overriding works.

I implemented in cryspy_crystal Atom and Atomset.
For Atomset the class Atom must be hashable.